\name{Infusion-internal}


% \alias{confintCDF.SLik}
\alias{addSimplexEq}
\alias{allCIs}
\alias{bounds1D}
\alias{calc_EI}
\alias{calc.lrthreshold.default}
\alias{calc.lrthreshold.SLik}
\alias{calc.lrthreshold.SLikp}
\alias{calc.lrthreshold}
\alias{check_logLs_input}
\alias{check_surface_input}
\alias{confintAll}
\alias{densityMixmod}
\alias{evalTotal}
\alias{findConstraints}
\alias{getVarPars}
\alias{get_best_clu_by_AIC}
\alias{greedyMAXMIN}
\alias{greedySelect}
\alias{infer_LR_pvalue_from_EDF}
\alias{infer_refDensity}
\alias{RMSEwrapper}
\alias{myrnorm} % because testthat appears to expect this from the package's namespace environment
\alias{neuralNet}
\alias{overcat}
\alias{par_RMSEwrapper}
\alias{par_wrapper}
\alias{predictWrap}
\alias{predict.dMixmod}
\alias{predict.SLik}
\alias{predict.SLik_j}
\alias{predict.MixmodResults}
\alias{predict.SLikp}
\alias{prettysignif}
\alias{print.SLik}
\alias{print.SLik_j}
\alias{print.logLs}
\alias{print.SLikp}
\alias{remove.misleading.replicates}
\alias{remove.pairswithNas}
\alias{rparam_from_SLik_j}
\alias{sample_vertices}
\alias{seekSimplex}
\alias{smoothEDF_s}
\alias{summary.logLs}
\alias{summary.SLik}
\alias{summary.SLik_j}
\alias{summary.SLikp}
\alias{SLR}
\alias{viridisOpts}

\alias{mapMM} % exported from spaMM
\alias{isPointInCHull} % exported from spaMM
\alias{resetCHull} % exported from spaMM


\title{Internal Infusion Functions}
\description{
  Internal Infusion functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
