\name{Infusion-internal}

\alias{projpath} % for local tests

\alias{calc.lrthreshold} % generic
\alias{calc.lrthreshold.default} % S3 methods
\alias{calc.lrthreshold.SLik}
\alias{calc.lrthreshold.SLikp}
\alias{predict.dMixmod}
\alias{predict.dMclust}
\alias{predict.SLik}
\alias{predict.MixmodResults}
\alias{predict.SLikp}
\alias{plot.MixmodResults}


% \alias{confintCDF.SLik}
%\alias{addSimplexEq}
%\alias{allCIs}
%\alias{calc_EI}
%\alias{check_logLs_input}
%\alias{check_surface_input}
%\alias{confintAll}
%\alias{densityMixmod}
%\alias{findConstraints}
%\alias{infer_LR_pvalue_from_EDF}
%\alias{infer_refDensity}
\alias{myrnorm} % (data!) because testthat appears to expect this from the package's namespace environment
%\alias{par_wrapper}
%\alias{predictWrap}
%\alias{sample_vertices}
%\alias{seekSimplex}
%\alias{smoothEDF_s}
%\alias{SLR}

\alias{mapMM} % exported from spaMM
\alias{isPointInCHull} % exported from blackbox
\alias{resetCHull} % exported from blackbox


\title{Internal Infusion Functions}
\description{
  Internal Infusion functions
}
\details{
  These are not to be called by the user, or are waiting for documentation to be written.
}
\keyword{ internal }
